feat: Optional data-available-fn to show no data msg
For cambridge non-corporate companies. Pass a function to Om shared that
takes a filter-spec and returns a boolean if data should be available.
In 3 chart call this function to see whether do show a message instead.
Show "Data not available" if not.

If shared fn not there then do old behaviour.